+++ sdk/decoder.go
@@ -8,5 +8,9 @@ import (
 //
 // This is only required if the chain supports decoding.
 type Decoder interface {
-	Decode(op types.Operation, abi string) (methodName string, args string, err error)
+	// Decode decodes the transaction data of a chain operation.
+	//
+	// contractInterfaces is the collection of functions and contract interfaces of the contract
+	// that the operation is interacting with.
+	Decode(op types.Transaction, contractInterfaces string) (DecodedOperation, error)
 }