+++ sdk/solana/common.go
@@ -45,9 +45,9 @@ func FindExpiringRootAndOpCountPDA(programID solana.PublicKey, pdaSeed PDASeed) 
 }
 
 func FindRootSignaturesPDA(
-	programID solana.PublicKey, msigID PDASeed, root common.Hash, validUntil uint32,
+	programID solana.PublicKey, msigID PDASeed, root common.Hash, validUntil uint32, authority solana.PublicKey,
 ) (solana.PublicKey, error) {
-	seeds := [][]byte{[]byte("root_signatures"), msigID[:], root[:], validUntilBytes(validUntil)}
+	seeds := [][]byte{[]byte("root_signatures"), msigID[:], root[:], validUntilBytes(validUntil), authority[:]}
 	return findPDA(programID, seeds)
 }
 
@@ -153,7 +153,18 @@ func sendAndConfirm(
 		return "", nil, fmt.Errorf("unable to validate and build instruction: %w", err)
 	}
 
-	result, err := solanaCommon.SendAndConfirm(ctx, client, []solana.Instruction{instruction}, auth, commitmentType)
+	return sendAndConfirmInstructions(ctx, client, auth, []solana.Instruction{instruction}, commitmentType)
+}
+
+// sendAndConfirm contains the default logic for sending and confirming instructions.
+func sendAndConfirmInstructions(
+	ctx context.Context,
+	client *rpc.Client,
+	auth solana.PrivateKey,
+	instructions []solana.Instruction,
+	commitmentType rpc.CommitmentType,
+) (string, *rpc.GetTransactionResult, error) {
+	result, err := solanaCommon.SendAndConfirm(ctx, client, instructions, auth, commitmentType)
 	if err != nil {
 		return "", nil, fmt.Errorf("unable to send instruction: %w", err)
 	}