+++ sdk/solana/configurer_test.go
@@ -6,10 +6,13 @@ import (
 	"testing"
 
 	"github.com/ethereum/go-ethereum/common"
+	ag_binary "github.com/gagliardetto/binary"
 	"github.com/gagliardetto/solana-go"
 	"github.com/gagliardetto/solana-go/rpc"
 	"github.com/google/go-cmp/cmp"
+	"github.com/google/go-cmp/cmp/cmpopts"
 	cselectors "github.com/smartcontractkit/chain-selectors"
+	bindings "github.com/smartcontractkit/chainlink-ccip/chains/solana/gobindings/mcm"
 	"github.com/stretchr/testify/require"
 
 	"github.com/smartcontractkit/mcms/sdk/solana/mocks"
@@ -23,9 +26,49 @@ func Test_NewConfigurer(t *testing.T) {
 	auth := solana.MustPrivateKeyFromBase58("DmPfeHBC8Brf8s5qQXi25bmJ996v6BHRtaLc6AH51yFGSqQpUMy1oHkbbXobPNBdgGH2F29PAmoq9ZZua4K9vCc")
 	chainSelector := types.ChainSelector(cselectors.SOLANA_DEVNET.Selector)
 
-	configurer := NewConfigurer(client, auth, chainSelector)
+	newAuthorityAccount := solana.NewWallet().PublicKey()
 
-	require.NotNil(t, configurer)
+	tests := []struct {
+		name                 string
+		constructorFn        func() *Configurer
+		wantSkipTransaction  bool
+		wantAuthorityAccount solana.PublicKey
+	}{
+		{
+			name: "standard args",
+			constructorFn: func() *Configurer {
+				return NewConfigurer(client, auth, chainSelector)
+			},
+			wantSkipTransaction:  false,
+			wantAuthorityAccount: auth.PublicKey(),
+		},
+		{
+			name: "skip transaction option",
+			constructorFn: func() *Configurer {
+				return NewConfigurer(client, auth, chainSelector, WithDoNotSendInstructionsOnChain())
+			},
+			wantSkipTransaction:  true,
+			wantAuthorityAccount: auth.PublicKey(),
+		},
+		{
+			name: "override authorityAccount",
+			constructorFn: func() *Configurer {
+				return NewConfigurer(client, auth, chainSelector, WithAuthorityAccount(newAuthorityAccount))
+			},
+			wantAuthorityAccount: newAuthorityAccount,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			configurer := tt.constructorFn()
+
+			require.NotNil(t, configurer)
+			require.Equal(t, tt.wantSkipTransaction, configurer.skipSend)
+			require.Equal(t, tt.wantAuthorityAccount, configurer.authorityAccount)
+		})
+	}
 }
 
 func TestConfigurer_SetConfig(t *testing.T) {
@@ -37,16 +80,21 @@ func TestConfigurer_SetConfig(t *testing.T) {
 	require.NoError(t, err)
 	defaultMcmConfig := &types.Config{Quorum: 1, Signers: []common.Address{common.HexToAddress("0x1")}}
 	clearRoot := false
+	newAuthorityAccount := solana.NewWallet().PublicKey()
 
 	tests := []struct {
-		name      string
-		mcmConfig *types.Config
-		setup     func(t *testing.T, configurer *Configurer, mockJSONRPCClient *mocks.JSONRPCClient)
-		want      string
-		wantErr   string
+		name             string
+		auth             solana.PrivateKey
+		options          []configurerOption
+		mcmConfig        *types.Config
+		setup            func(t *testing.T, configurer *Configurer, mockJSONRPCClient *mocks.JSONRPCClient)
+		wantHash         string
+		wantInstructions []solana.Instruction
+		wantErr          string
 	}{
 		{
-			name:      "success",
+			name:      "success - send instructions",
+			auth:      auth,
 			mcmConfig: defaultMcmConfig,
 			setup: func(t *testing.T, configurer *Configurer, mockJSONRPCClient *mocks.JSONRPCClient) {
 				t.Helper()
@@ -62,16 +110,52 @@ func TestConfigurer_SetConfig(t *testing.T) {
 				mockSolanaTransaction(t, mockJSONRPCClient, 13, 23,
 					"52f3VmvW7m9uTQu3PtyibgxnAvEuXDmm9umuHherGjS4pzRR7QXRDKnZhh6b95P7pQxzTgvE1muMNKYEY7YWsS3G", nil, nil)
 			},
-			want: "52f3VmvW7m9uTQu3PtyibgxnAvEuXDmm9umuHherGjS4pzRR7QXRDKnZhh6b95P7pQxzTgvE1muMNKYEY7YWsS3G",
+			wantHash: "52f3VmvW7m9uTQu3PtyibgxnAvEuXDmm9umuHherGjS4pzRR7QXRDKnZhh6b95P7pQxzTgvE1muMNKYEY7YWsS3G",
+			wantInstructions: []solana.Instruction{
+				bindings.NewInitSignersInstructionBuilder().Build(),
+				bindings.NewAppendSignersInstructionBuilder().Build(),
+				bindings.NewFinalizeSignersInstructionBuilder().Build(),
+				bindings.NewSetConfigInstructionBuilder().Build(),
+			},
+		},
+		{
+			name:      "success - do not send instructions",
+			auth:      auth,
+			options:   []configurerOption{WithDoNotSendInstructionsOnChain()},
+			mcmConfig: defaultMcmConfig,
+			setup:     func(t *testing.T, configurer *Configurer, mockJSONRPCClient *mocks.JSONRPCClient) { t.Helper() },
+			wantHash:  "",
+			wantInstructions: []solana.Instruction{
+				bindings.NewInitSignersInstructionBuilder().Build(),
+				bindings.NewAppendSignersInstructionBuilder().Build(),
+				bindings.NewFinalizeSignersInstructionBuilder().Build(),
+				bindings.NewSetConfigInstructionBuilder().Build(),
+			},
+		},
+		{
+			name:      "success - override authorityAccount acciybt",
+			auth:      auth,
+			options:   []configurerOption{WithDoNotSendInstructionsOnChain(), WithAuthorityAccount(newAuthorityAccount)},
+			mcmConfig: defaultMcmConfig,
+			setup:     func(t *testing.T, configurer *Configurer, mockJSONRPCClient *mocks.JSONRPCClient) { t.Helper() },
+			wantHash:  "",
+			wantInstructions: []solana.Instruction{
+				bindings.NewInitSignersInstructionBuilder().Build(),
+				bindings.NewAppendSignersInstructionBuilder().Build(),
+				bindings.NewFinalizeSignersInstructionBuilder().Build(),
+				bindings.NewSetConfigInstructionBuilder().Build(),
+			},
 		},
 		{
 			name:      "failure: too many signers",
+			auth:      auth,
 			mcmConfig: &types.Config{Quorum: 1, Signers: generateSigners(t, 181)},
 			setup:     func(t *testing.T, configurer *Configurer, mockJSONRPCClient *mocks.JSONRPCClient) { t.Helper() },
 			wantErr:   "too many signers (max 180)",
 		},
 		{
 			name:      "failure: initialize signers error",
+			auth:      auth,
 			mcmConfig: defaultMcmConfig,
 			setup: func(t *testing.T, configurer *Configurer, mockJSONRPCClient *mocks.JSONRPCClient) {
 				t.Helper()
@@ -80,10 +164,11 @@ func TestConfigurer_SetConfig(t *testing.T) {
 					"4PQcRHQJT4cRQZooAhZMAP9ZXJsAka9DeKvXeYvXAvPpHb4Qkc5rmTSHDA2SZSh9aKPBguBx4kmcyHHbkytoAiRr",
 					nil, fmt.Errorf("initialize signers error"))
 			},
-			wantErr: "unable to initialize signers: unable to send instruction: initialize signers error",
+			wantErr: "unable to set config: unable to send instruction 0 - initSigners: unable to send instruction: initialize signers error",
 		},
 		{
 			name:      "failure: append signers error",
+			auth:      auth,
 			mcmConfig: defaultMcmConfig,
 			setup: func(t *testing.T, configurer *Configurer, mockJSONRPCClient *mocks.JSONRPCClient) {
 				t.Helper()
@@ -97,10 +182,11 @@ func TestConfigurer_SetConfig(t *testing.T) {
 					"7D9XEYRnCn1D5JFrrYMPUaHfog7Vnj5rbPdj7kbULa4hKq7GsnA7Q8KNQfLEgfCawBsW4dcH2MQAp4km1dnjr6V",
 					nil, fmt.Errorf("append signers error"))
 			},
-			wantErr: "unable to append signers (0): unable to send instruction: append signers error",
+			wantErr: "unable to set config: unable to send instruction 1 - appendSigners0: unable to send instruction: append signers error",
 		},
 		{
 			name:      "failure: finalize signers error",
+			auth:      auth,
 			mcmConfig: defaultMcmConfig,
 			setup: func(t *testing.T, configurer *Configurer, mockJSONRPCClient *mocks.JSONRPCClient) {
 				t.Helper()
@@ -116,10 +202,11 @@ func TestConfigurer_SetConfig(t *testing.T) {
 					"2iEeniu3QUgXNsjau8r7fZ7XLb2g1F3q9VJJKvRyyFz4hHgVvhGkLgSUdmRumfXKWv8spJ9ihudGFyPZsPGdp4Ya",
 					nil, fmt.Errorf("finalize signers error"))
 			},
-			wantErr: "unable to finalize signers: unable to send instruction: finalize signers error",
+			wantErr: "unable to set config: unable to send instruction 2 - finalizeSigners: unable to send instruction: finalize signers error",
 		},
 		{
 			name:      "failure: set config error",
+			auth:      auth,
 			mcmConfig: &types.Config{Quorum: 1, Signers: []common.Address{common.HexToAddress("0x1")}},
 			setup: func(t *testing.T, configurer *Configurer, mockJSONRPCClient *mocks.JSONRPCClient) {
 				t.Helper()
@@ -137,21 +224,23 @@ func TestConfigurer_SetConfig(t *testing.T) {
 					"52f3VmvW7m9uTQu3PtyibgxnAvEuXDmm9umuHherGjS4pzRR7QXRDKnZhh6b95P7pQxzTgvE1muMNKYEY7YWsS3G",
 					nil, fmt.Errorf("set config error"))
 			},
-			wantErr: "unable to set config: unable to send instruction: set config error",
+			wantErr: "unable to set config: unable to send instruction 3 - setConfig: unable to send instruction: set config error",
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 
-			configurer, mockJSONRPCClient := newTestConfigurer(t, auth, chainSelector)
+			configurer, mockJSONRPCClient := newTestConfigurer(t, tt.auth, chainSelector, tt.options...)
 			tt.setup(t, configurer, mockJSONRPCClient)
 
 			got, err := configurer.SetConfig(ctx, ContractAddress(testMCMProgramID, testPDASeed), tt.mcmConfig, clearRoot)
 
 			if tt.wantErr == "" {
 				require.NoError(t, err)
-				require.Empty(t, cmp.Diff(tt.want, got.Hash))
+				require.Empty(t, cmp.Diff(tt.wantHash, got.Hash))
+				require.Empty(t, cmp.Diff(tt.wantInstructions, got.RawData.([]solana.Instruction),
+					cmpopts.IgnoreFields(ag_binary.BaseVariant{}, "Impl")))
 			} else {
 				require.ErrorContains(t, err, tt.wantErr)
 			}
@@ -161,11 +250,13 @@ func TestConfigurer_SetConfig(t *testing.T) {
 
 // ----- helpers -----
 
-func newTestConfigurer(t *testing.T, auth solana.PrivateKey, chainSelector types.ChainSelector) (*Configurer, *mocks.JSONRPCClient) {
+func newTestConfigurer(
+	t *testing.T, auth solana.PrivateKey, chainSelector types.ChainSelector, options ...configurerOption,
+) (*Configurer, *mocks.JSONRPCClient) {
 	t.Helper()
 
 	mockJSONRPCClient := mocks.NewJSONRPCClient(t)
 	client := rpc.NewWithCustomRPCClient(mockJSONRPCClient)
 
-	return NewConfigurer(client, auth, chainSelector), mockJSONRPCClient
+	return NewConfigurer(client, auth, chainSelector, options...), mockJSONRPCClient
 }