+++ timelock_executable_test.go
@@ -12,8 +12,6 @@ import (
 	"github.com/ethereum/go-ethereum/common"
 	geth_types "github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/gagliardetto/solana-go"
-	"github.com/gagliardetto/solana-go/rpc"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 
@@ -26,7 +24,6 @@ import (
 	"github.com/smartcontractkit/mcms/sdk/evm"
 	"github.com/smartcontractkit/mcms/sdk/evm/bindings"
 	"github.com/smartcontractkit/mcms/sdk/mocks"
-	solanasdk "github.com/smartcontractkit/mcms/sdk/solana"
 	"github.com/smartcontractkit/mcms/types"
 )
 
@@ -41,6 +38,8 @@ func Test_NewTimelockExecutable(t *testing.T) {
 	t.Parallel()
 
 	var (
+		ctx = context.Background()
+
 		executor = mocks.NewTimelockExecutor(t)
 
 		chainMetadata = map[types.ChainSelector]types.ChainMetadata{
@@ -99,10 +98,56 @@ func Test_NewTimelockExecutable(t *testing.T) {
 				Operations:        batchOps,
 			},
 			giveExecutors: map[types.ChainSelector]sdk.TimelockExecutor{
-				types.ChainSelector(1): executor,
+				chaintest.Chain1Selector: executor,
 			},
 			wantErr:    false,
 			wantErrMsg: "",
+		},
+		{
+			name: "failure: converter from executor error",
+			giveProposal: &TimelockProposal{
+				BaseProposal: BaseProposal{
+					Version:              "v1",
+					Kind:                 types.KindTimelockProposal,
+					Description:          "description",
+					ValidUntil:           2004259681,
+					OverridePreviousRoot: false,
+					Signatures:           []types.Signature{},
+					ChainMetadata:        chainMetadata,
+				},
+				Action:            types.TimelockActionSchedule,
+				Delay:             types.MustParseDuration("1h"),
+				TimelockAddresses: timelockAddresses,
+				Operations:        batchOps,
+			},
+			giveExecutors: map[types.ChainSelector]sdk.TimelockExecutor{
+				types.ChainSelector(1): executor,
+			},
+			wantErr:    true,
+			wantErrMsg: "unable to set predecessors: unable to create converter from executor: chain family not found for selector",
+		},
+		{
+			name: "failure: convert error",
+			giveProposal: &TimelockProposal{
+				BaseProposal: BaseProposal{
+					Version:              "v1",
+					Kind:                 types.KindTimelockProposal,
+					Description:          "description",
+					ValidUntil:           2004259681,
+					OverridePreviousRoot: false,
+					Signatures:           []types.Signature{},
+					ChainMetadata:        chainMetadata,
+				},
+				Action:            types.TimelockActionSchedule,
+				Delay:             types.MustParseDuration("1h"),
+				TimelockAddresses: timelockAddresses,
+				Operations:        batchOps,
+			},
+			giveExecutors: map[types.ChainSelector]sdk.TimelockExecutor{
+				chaintest.Chain2Selector: executor,
+			},
+			wantErr:    true,
+			wantErrMsg: "unable to set predecessors: unable to find converter for chain selector",
 		},
 	}
 
@@ -110,7 +155,7 @@ func Test_NewTimelockExecutable(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
 
-			_, err := NewTimelockExecutable(tt.giveProposal, tt.giveExecutors)
+			_, err := NewTimelockExecutable(ctx, tt.giveProposal, tt.giveExecutors)
 
 			if tt.wantErr {
 				require.Error(t, err)
@@ -206,34 +251,6 @@ func Test_TimelockExecutable_Execute(t *testing.T) {
 			want:   "signature",
 		},
 		{
-			name: "failure: converter from executor error",
-			setup: func(t *testing.T) (*TimelockProposal, map[types.ChainSelector]sdk.TimelockExecutor) {
-				t.Helper()
-				executors := map[types.ChainSelector]sdk.TimelockExecutor{
-					types.ChainSelector(12345789): mocks.NewTimelockExecutor(t),
-				}
-
-				return defaultProposal(), executors
-			},
-			wantErr: "unable to set predecessors: unable to create converter from executor: chain family not found for selector",
-		},
-		{
-			name: "failure: convert error",
-			setup: func(t *testing.T) (*TimelockProposal, map[types.ChainSelector]sdk.TimelockExecutor) {
-				t.Helper()
-
-				solanaClient := &rpc.Client{}
-				solanaAuth, err := solana.NewRandomPrivateKey()
-				require.NoError(t, err)
-				executors := map[types.ChainSelector]sdk.TimelockExecutor{
-					chaintest.Chain4Selector: solanasdk.NewTimelockExecutor(solanaClient, solanaAuth),
-				}
-
-				return defaultProposal(), executors
-			},
-			wantErr: "unable to set predecessors: unable to find converter for chain selector",
-		},
-		{
 			name: "failure: execute error",
 			setup: func(t *testing.T) (*TimelockProposal, map[types.ChainSelector]sdk.TimelockExecutor) {
 				t.Helper()
@@ -254,7 +271,7 @@ func Test_TimelockExecutable_Execute(t *testing.T) {
 			t.Parallel()
 
 			proposal, executors := tt.setup(t)
-			timelockExecutable, err := NewTimelockExecutable(proposal, executors)
+			timelockExecutable, err := NewTimelockExecutable(ctx, proposal, executors)
 			require.NoError(t, err)
 
 			var got types.TransactionResult
@@ -523,7 +540,7 @@ func scheduleAndExecuteGrantRolesProposal(t *testing.T, ctx context.Context, tar
 	}
 
 	// Create new executable
-	tExecutable, err := NewTimelockExecutable(&proposal, tExecutors)
+	tExecutable, err := NewTimelockExecutable(ctx, &proposal, tExecutors)
 	require.NoError(t, err)
 
 	for i := range predecessors {
@@ -556,27 +573,44 @@ func scheduleAndExecuteGrantRolesProposal(t *testing.T, ctx context.Context, tar
 	require.NoError(t, sim.Backend.AdjustTime(5*time.Second))
 	sim.Backend.Commit() // Note < 1.14 geth needs a commit after adjusting time.
 
-	// Check that the operation is now ready
+	opIdx := 0
+
+	// IsReady
 	err = tExecutable.IsReady(ctx)
 	require.NoError(t, err)
-
-	// Check IsChainReady function succeeds
 	for chainSelector := range proposal.ChainMetadata {
 		err = tExecutable.IsChainReady(ctx, chainSelector)
 		require.NoError(t, err)
 	}
 
+	// !IsDone
+	err = tExecutable.IsOperationDone(ctx, opIdx)
+	require.Error(t, err)
+	for chainSelector := range proposal.ChainMetadata {
+		err = tExecutable.IsChainDone(ctx, chainSelector)
+		require.Error(t, err)
+	}
+
 	// Execute the proposal
-	idx := 0
-	_, err = tExecutable.Execute(ctx, idx)
+	_, err = tExecutable.Execute(ctx, opIdx)
 	require.NoError(t, err)
 	sim.Backend.Commit()
-	opID, err := tExecutable.GetOpID(ctx, idx, proposal.Operations[idx], proposal.Operations[idx].ChainSelector)
+
+	// IsDone
+	err = tExecutable.IsOperationDone(ctx, opIdx)
 	require.NoError(t, err)
-	// Check that the operation is done
-	isOperationDone, err := timelockC.IsOperationDone(&bind.CallOpts{}, opID)
-	require.NoError(t, err)
-	require.True(t, isOperationDone)
+	for chainSelector := range proposal.ChainMetadata {
+		err = tExecutable.IsChainDone(ctx, chainSelector)
+		require.NoError(t, err)
+	}
+
+	// !IsReady
+	err = tExecutable.IsOperationReady(ctx, opIdx)
+	require.Error(t, err)
+	for chainSelector := range proposal.ChainMetadata {
+		err = tExecutable.IsChainReady(ctx, chainSelector)
+		require.Error(t, err)
+	}
 
 	// Check the state of the timelock contract
 	for _, role := range targetRoles {
@@ -683,7 +717,7 @@ func scheduleAndCancelGrantRolesProposal(t *testing.T, ctx context.Context, targ
 	}
 
 	// Create new executable
-	tExecutable, err := NewTimelockExecutable(&proposal, tExecutors)
+	tExecutable, err := NewTimelockExecutable(ctx, &proposal, tExecutors)
 	require.NoError(t, err)
 
 	for i := range predecessors {
@@ -813,6 +847,8 @@ func scheduleAndCancelGrantRolesProposal(t *testing.T, ctx context.Context, targ
 func Test_TimelockExecutable_GetChainSpecificIndex(t *testing.T) {
 	t.Parallel()
 
+	ctx := context.Background()
+
 	// We'll create a helper function for generating a minimal proposal
 	// that won't fail validation. It includes multiple operations across
 	// different chain selectors.
@@ -898,7 +934,7 @@ func Test_TimelockExecutable_GetChainSpecificIndex(t *testing.T) {
 		executors := map[types.ChainSelector]sdk.TimelockExecutor{}
 
 		// Create TimelockExecutable
-		tlExecutable, err := NewTimelockExecutable(proposal, executors)
+		tlExecutable, err := NewTimelockExecutable(ctx, proposal, executors)
 		require.NoError(t, err)
 
 		// Each test-case checks the chain-specific index for a given global index